fix: rename routes and create navigation settings for requests
